requirements3.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.5
django-extensions==2.2.1
django-filter==2.2.0
django-rest-swagger==0.3.10
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-tweaks==0.9.2
elasticsearch==7.0.4
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
importlib-metadata==0.21
kombu==4.6.4
more-itertools==7.2.0
mysqlclient==1.4.4
Pillow==6.1.0
PyJWT==1.7.1
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
reportlab==3.5.23
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
webencodings==0.5.1
xhtml2pdf==0.2.3
zipp==0.6.0
